+++ dashboard/src/routes.tsx
@@ -1,9 +1,9 @@
 import App from "App";
+import AdminGuard from "pages/admin/AdminGuard";
 import AdminPage from "pages/admin/AdminPage";
-import AdminGuard from "pages/AdminGuard";
-import AdminServerPage from "pages/AdminServerPage";
-import AdminServersPage from "pages/AdminServersPage";
-import AdminSubscriptionsPage from "pages/AdminSubscriptionsPage";
+import AdminServerPage from "pages/admin/servers/AdminServerPage";
+import AdminServersPage from "pages/admin/servers/AdminServersPage";
+import AdminSubscriptionsPage from "pages/admin/subscriptions/AdminSubscriptionsPage";
 import Auth from "pages/Auth";
 import AuthGuard from "pages/AuthGuard";
 import BattlesPage from "pages/BattlesPage";
